# doc/swagger.yml
openapi: 3.0.1
info:
  title: dating server API
  description: Описание методов API сервера мобильного приложения
  version: 0.24.6
servers:
  - url: https://dev.dating.eagledev.ru/api/v1
  - url: https://prod.dating.eagledev.ru/api/v1
  - url: http://localhost:8032/api/v1/
paths:
...
  /auth/request-code:
    post:
      tags:
        - Авторизация
      summary: Запрос кода
      description: Запрос одноразового кода для входа или регистрации
      requestBody:
        content:
          application/json:
            schema:
              required:
                - email
                - hash
              type: object
              properties:
                email:
                  type: string
                  description: email
                  example: mail@example.com
                hash:
                  type: string
                  description: секретный хэш для защиты от спама sha256(email+соль)
                  example: 123abc
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех операции
                    example: true
        "400":
          description: некорректный формат запроса/ошибка в теле запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "429":
          description: повторный запрос, код уже был отправлен (code already sent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: server error
          content: { }
      x-codegen-request-body-name: data
  /auth/login:
    post:
      tags:
        - Авторизация
      summary: Вход или регистрация
      description: "Вход в систему, если пользователя не существовало, то он будет\
        \ сразу создан"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        "200":
          description: успешная авторизация
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: request success
                    example: true
                  data:
                    required:
                      - expire
                      - isNewUser
                      - refreshToken
                      - token
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT-токен для авторизации запросов
                        example: string-jwt-token
                      refreshToken:
                        type: string
                        description: рефреш-токен для обновления JWT-токена
                        example: test-refresh-token
                      expire:
                        type: integer
                        description: время жизни токена в секундах
                        example: 3600
                      isNewUser:
                        type: boolean
                        description: статус пользователя, новый или нет
                        example: true
        "400":
          description: некорректный формат запроса/ошибка в теле запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: неверный логин или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codegen-request-body-name: data
  /auth/refresh-token:
    post:
      tags:
        - Авторизация
      summary: Обновление токена
      description: получение нового токена с помощью рефреш-токена
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: request success
                    example: true
                  data:
                    required:
                      - expire
                      - refreshToken
                      - token
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT-токен для авторизации запросов
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxMjh9.Us_MSJotZOGcunjSAy0_CT-pt3asJw-8d1srPOYRSlk
                      refreshToken:
                        type: string
                        description: рефреш-токен для обновления JWT-токена
                        example: test-refresh-token
                      expire:
                        type: integer
                        description: время жизни токена в секундах
                        example: 3600
        "400":
          description: некорректный формат запроса/ошибка в теле запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: невалидный рефреш-токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codegen-request-body-name: data
  /auth/vk:
    post:
      tags:
        - Авторизация
      summary: вход через ВК
      description: вход через ВК
      requestBody:
        content:
          application/json:
            schema:
              required:
                - accessToken
              properties:
                accessToken:
                  type: string
                  description: токен, полученный после авторизации чреез VK ID
        required: true
      responses:
        200:
          description: OK - успешная авторизация, вернется токен и рефреш токен
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: request success
                    example: true
                  data:
                    required:
                      - expire
                      - refreshToken
                      - token
                      - isNewUser
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT-токен для авторизации запросов
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxMjh9.Us_MSJotZOGcunjSAy0_CT-pt3asJw-8d1srPOYRSlk
                      refreshToken:
                        type: string
                        description: рефреш-токен для обновления JWT-токена
                        example: test-refresh-token
                      expire:
                        type: integer
                        description: время жизни токена в секундах
                        example: 3600
                      isNewUser:
                        type: boolean
                        description: статус пользователя, новый или нет
                        example: true
        400:
          description: todo
        500:
          description: internal server error
...
  /chat:
    get:
      tags:
        - Чат
      summary: Список чатов
      description: todo
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: список мэтчей с открытыми чатам
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Chat'
        "500":
          description: server error
          content: { }
  /chat/messages:
    post:
      tags:
        - Чат
      summary: Список сообщений в чате
      description: todo
      security:
        - bearerAuth: [ ]
      requestBody:
        content:
          application/json:
            schema:
              required:
                - limit
                - matchID
              type: object
              properties:
                matchID:
                  type: string
                  description: match (chat) uuid
                limit:
                  type: integer
                offset:
                  type: integer
        required: true
      responses:
        "200":
          description: массив сообщений
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        "500":
          description: server error
          content: { }
      x-codegen-request-body-name: data
  /chat/send-message:
    post:
      tags:
        - Чат
      summary: Отправить сообщение в чат
      security:
        - bearerAuth: [ ]
      description: Отправить сообщение в чат мэтча
      requestBody:
        content:
          application/json:
            schema:
              required:
                - matchID
                - text
              type: object
              properties:
                matchID:
                  type: string
                  description: "match (chat) uuid"
                text:
                  type: string
                  description: "текст сообщения"
        required: true
      responses:
        200:
          description: OK, сообщение отправлено
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
        404:
          description: Несуществующий чат\мэтч
        400:
          description: неверный запрос
        403:
          description: Нет доступа
        500:
          description: server error
...
components:
  schemas:
    LoginRequest:
      required:
        - code
        - email
      type: object
      properties:
        email:
          type: string
          description: email
          example: mail@example.com
        code:
          type: string
          description: однаразовый код
          example: "1234"
...
    RefreshTokenRequest:
      required:
        - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
          description: refresh-токен
          example: test-refresh-token
...
    ErrorResponse:
      required:
        - error
        - success
      type: object
      properties:
        success:
          type: boolean
          description: request success
          example: false
        error:
          required:
            - code
            - message
          type: object
          properties:
            code:
              type: integer
              description: internal error code
              example: 1001
            message:
              type: string
              description: error text
              example: example error
    UserProfile:
      required:
        - ID
      type: object
      properties:
        ID:
          type: string
          description: uuid пользователя
        name:
          type: string
          description: имя
        sex:
          type: string
          description: пол
        birthday:
          type: string
          description: дата рождения
        city:
          type: string
          description: город
        about:
          type: string
          description: о себе
        photos:
          type: array
          items:
            $ref: '#/components/schemas/UserPhoto'
        distance:
          type: integer
          description: расстояние до пользователя в километрах
          example: 4389
    UserPhoto:
      required:
        - ID
        - URL
        - isMain
      type: object
      properties:
        ID:
          type: string
          description: uuid фотографии
          example: 6a7485ca-2020-470d-b0b5-04f3ff19f12f
        URL:
          type: string
          description: URL адрес картинки
          example: https://example.example/example.jpg
        isMain:
          type: boolean
          description: флаг основное фото профиля
          example: false
        createdAt:
          type: string
          description: дата добавления
...
    Message:
      required:
        - createdAt
        - delivered
        - id
        - isRead
        - text
        - userId
      type: object
      properties:
        id:
          type: string
          description: uuid сообщения
          example: uuid
        userId:
          type: string
          description: "uuid пользователя, отправившего сообщение"
          example: uuid
        text:
          type: string
          description: текст сообщения
          example: "hello, world"
        isRead:
          type: boolean
          description: статус просмотра сообщения
          example: true
        delivered:
          type: boolean
          description: статус доставки
          example: true
        createdAt:
          type: integer
          description: время создания сообщения
          example: 123456778
    Chat:
      required:
        - ID
        - lastMessage
        - profile
      type: object
      properties:
        ID:
          type: string
          description: match uuid
          example: uuid
        profile:
          $ref: '#/components/schemas/UserProfile'
        lastMessage:
          $ref: '#/components/schemas/Message'
        createdAt:
          type: string
          description: дата создания мэтча
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT